+++ lib/modules/RateLimit/DurableObject.ts
@@ -1,31 +1,16 @@
+import {isFn} from '@valkyriestudios/utils/function';
 import {type TriFrostCFDurableObjectNamespace} from '../../types/providers';
 import {type LazyInitFn} from '../../utils/Lazy';
 import {DurableObjectStore} from '../_storage/DurableObject';
-import {
-    TriFrostRateLimit,
-    type TriFrostRateLimitStrategy,
-    type TriFrostRateLimitKeyGenerator,
-    type TriFrostRateLimitExceededFunction,
-} from './_RateLimit';
+import {TriFrostRateLimit, type TriFrostRateLimitOptions} from './_RateLimit';
 
 export class DurableObjectRateLimit <Env extends Record<string, any> = Record<string, any>> extends TriFrostRateLimit<Env> {
 
-    constructor (cfg: {
-		window?: number;
-		strategy?: TriFrostRateLimitStrategy;
-		keygen?: TriFrostRateLimitKeyGenerator;
-		headers?: boolean;
-		exceeded?: TriFrostRateLimitExceededFunction;
-		store: LazyInitFn<TriFrostCFDurableObjectNamespace, Env>;
-	}) {
-        super({
-            window: cfg.window ?? 60_000,
-            strategy: cfg.strategy ?? 'fixed',
-            keygen: cfg.keygen,
-            headers: cfg.headers,
-            exceeded: cfg.exceeded,
-            store: ({env}) => new DurableObjectStore(cfg.store({env}), 'ratelimit'),
-        });
+    constructor (cfg: Omit<TriFrostRateLimitOptions<Env>, 'store'> & {
+        store: LazyInitFn<TriFrostCFDurableObjectNamespace, Env>
+    }) {
+        if (!isFn(cfg?.store)) throw new Error('DurableObjectRateLimit: Expected a store initializer');
+        super({...cfg, store: ({env}) => new DurableObjectStore(cfg.store({env}), 'ratelimit')});
     }
 
 }