+++ lib/modules/RateLimit/KV.ts
@@ -1,31 +1,16 @@
+import {isFn} from '@valkyriestudios/utils/function';
 import {type TriFrostCFKVNamespace} from '../../types/providers';
 import {type LazyInitFn} from '../../utils/Lazy';
 import {KVStore} from '../_storage/KV';
-import {
-    TriFrostRateLimit,
-    type TriFrostRateLimitStrategy,
-    type TriFrostRateLimitKeyGenerator,
-    type TriFrostRateLimitExceededFunction,
-} from './_RateLimit';
+import {TriFrostRateLimit, type TriFrostRateLimitOptions} from './_RateLimit';
 
 export class KVRateLimit <Env extends Record<string, any> = Record<string, any>> extends TriFrostRateLimit<Env> {
 
-    constructor (cfg: {
-        window?: number;
-        strategy?: TriFrostRateLimitStrategy;
-        keygen?: TriFrostRateLimitKeyGenerator;
-        headers?: boolean;
-        exceeded?: TriFrostRateLimitExceededFunction;
-        store: LazyInitFn<TriFrostCFKVNamespace, Env>;
+    constructor (cfg: Omit<TriFrostRateLimitOptions<Env>, 'store'> & {
+        store: LazyInitFn<TriFrostCFKVNamespace, Env>
     }) {
-        super({
-            window: cfg.window ?? 60_000,
-            strategy: cfg.strategy ?? 'fixed',
-            keygen: cfg.keygen,
-            headers: cfg.headers,
-            exceeded: cfg.exceeded,
-            store: ({env}) => new KVStore(cfg.store({env})),
-        });
+        if (!isFn(cfg?.store)) throw new Error('KVRateLimit: Expected a store initializer');
+        super({...cfg, store: ({env}) => new KVStore(cfg.store({env}))});
     }
 
 }