+++ lib/modules/RateLimit/Memory.ts
@@ -1,26 +1,16 @@
 import {isNeArray} from '@valkyriestudios/utils/array';
+import {isIntGt} from '@valkyriestudios/utils/number';
 import {type LazyInitFn} from '../../utils/Lazy';
 import {MemoryStore} from '../_storage/Memory';
 import {type TriFrostRateLimitObject} from './strategies/_Strategy';
-import {
-    TriFrostRateLimit,
-    type TriFrostRateLimitStrategy,
-    type TriFrostRateLimitKeyGenerator,
-    type TriFrostRateLimitExceededFunction,
-} from './_RateLimit';
+import {TriFrostRateLimit, type TriFrostRateLimitOptions} from './_RateLimit';
 
 export class MemoryRateLimit<Env extends Record<string, any> = Record<string, any>> extends TriFrostRateLimit<Env> {
 
-    constructor (cfg: {
-		window?: number;
-		strategy?: TriFrostRateLimitStrategy;
-		keygen?: TriFrostRateLimitKeyGenerator;
-		headers?: boolean;
-		exceeded?: TriFrostRateLimitExceededFunction;
-	}) {
-	    const window = cfg.window ?? 60_000;
+    constructor (cfg?: Omit<TriFrostRateLimitOptions<Env>, 'store'>) {
+        const window = isIntGt(cfg?.window, 0) ? cfg.window : 60_000;
 
-        const store = cfg.strategy === 'sliding'
+        const store = cfg?.strategy === 'sliding'
             ? new MemoryStore<number[]>({
                 gc_interval: 60_000,
                 gc_filter: (_, timestamps, now) => isNeArray(timestamps) && timestamps[timestamps.length - 1] < (now - window),
@@ -31,11 +21,7 @@ export class MemoryRateLimit<Env extends Record<string, any> = Record<string, an
             });
 
         super({
-            window,
-            strategy: cfg.strategy ?? 'fixed',
-            keygen: cfg.keygen,
-            headers: cfg.headers,
-            exceeded: cfg.exceeded,
+            ...cfg || {},
             store: (() => store) as LazyInitFn<MemoryStore, Env>,
         });
     }